pyproject.toml
[project]
name = "sf-automation-switch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "lxml==5.3.0",
    "marshmallow-enum>=1.5.1",
    "marshmallow>=3.23.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.0",
    "simple-salesforce>=1.12.6",
]
